+++ src/Factory.php
@@ -5,18 +5,18 @@ namespace Trukes\ThreadsApiPhpClient;
 
 use Http\Discovery\Psr18ClientDiscovery;
 use Psr\Http\Client\ClientInterface;
-use Trukes\ThreadsApiPhpClient\DTO\Config;
-use Trukes\ThreadsApiPhpClient\DTO\Transporter\AccessToken;
-use Trukes\ThreadsApiPhpClient\DTO\Transporter\BaseUri;
-use Trukes\ThreadsApiPhpClient\DTO\Transporter\BodyForm;
-use Trukes\ThreadsApiPhpClient\DTO\Transporter\Headers;
-use Trukes\ThreadsApiPhpClient\DTO\Transporter\QueryParams;
 use Trukes\ThreadsApiPhpClient\Reference\Container\Insights\Insights;
 use Trukes\ThreadsApiPhpClient\Reference\Container\Media\Media;
 use Trukes\ThreadsApiPhpClient\Reference\Container\Publish\Publish;
 use Trukes\ThreadsApiPhpClient\Reference\Container\ReplyManagement\ReplyManagement;
 use Trukes\ThreadsApiPhpClient\Reference\Container\User\User;
-use Trukes\ThreadsApiPhpClient\Reference\Reference;
+use Trukes\ThreadsApiPhpClient\Transporter\Transporter;
+use Trukes\ThreadsApiPhpClient\Transporter\ValueObject\AccessToken;
+use Trukes\ThreadsApiPhpClient\Transporter\ValueObject\BaseUri;
+use Trukes\ThreadsApiPhpClient\Transporter\ValueObject\BodyForm;
+use Trukes\ThreadsApiPhpClient\Transporter\ValueObject\Headers;
+use Trukes\ThreadsApiPhpClient\Transporter\ValueObject\QueryParams;
+use Trukes\ThreadsApiPhpClient\ValueObject\Config;
 
 final class Factory
 {
@@ -45,7 +45,7 @@ final class Factory
         return $this;
     }
 
-    public function make(): Reference
+    public function make(): Client
     {
         $headers = Headers::create();
 
@@ -61,7 +61,7 @@ final class Factory
 
         $transporter = new Transporter($client, $baseUri, $headers, $queryParams, $bodyForm, $accessToken);
 
-        return new Reference(
+        return new Client(
             new Publish($transporter),
             new Media($transporter),
             new ReplyManagement($transporter),
