+++ src/Reference/Container/User/User.php
@@ -3,14 +3,14 @@ declare(strict_types=1);
 
 namespace Trukes\ThreadsApiPhpClient\Reference\Container\User;
 
-use Trukes\ThreadsApiPhpClient\DTO\Payload;
-use Trukes\ThreadsApiPhpClient\DTO\Response;
 use Trukes\ThreadsApiPhpClient\Reference\Container\User\Factory\Profile;
 use Trukes\ThreadsApiPhpClient\Reference\Container\User\Factory\PublishLimit;
 use Trukes\ThreadsApiPhpClient\Reference\Container\User\Factory\Replies;
 use Trukes\ThreadsApiPhpClient\Reference\Container\User\Factory\Threads;
-use Trukes\ThreadsApiPhpClient\TransporterInterface;
+use Trukes\ThreadsApiPhpClient\Transporter\TransporterInterface;
 use Trukes\ThreadsApiPhpClient\TransporterTrait;
+use Trukes\ThreadsApiPhpClient\ValueObject\Payload;
+use Trukes\ThreadsApiPhpClient\ValueObject\Response;
 
 final class User implements UserInterface
 {
