+++ src/Threads.php
@@ -1,12 +1,11 @@
 <?php
+declare(strict_types=1);
 
 namespace Trukes\ThreadsApiPhpClient;
 
-use Trukes\ThreadsApiPhpClient\Reference\Reference;
-
 final class Threads
 {
-    public static function client(?string $accessToken): Reference
+    public static function client(?string $accessToken): Client
     {
         return self::factory()
             ->withConfig()
