+++ src/TransporterTrait.php
@@ -2,6 +2,8 @@
 
 namespace Trukes\ThreadsApiPhpClient;
 
+use Trukes\ThreadsApiPhpClient\Transporter\TransporterInterface;
+
 trait TransporterTrait
 {
     public function __construct(private readonly TransporterInterface $transporter)
