+++ tests/ThreadsTest.php
@@ -5,21 +5,21 @@ namespace Tests;
 
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Client\ClientInterface;
-use Trukes\ThreadsApiPhpClient\DTO\Config;
-use Trukes\ThreadsApiPhpClient\Reference\Reference;
+use Trukes\ThreadsApiPhpClient\Client;
 use Trukes\ThreadsApiPhpClient\Threads;
+use Trukes\ThreadsApiPhpClient\ValueObject\Config;
 
 final class ThreadsTest extends TestCase
 {
     public function testThreadsClient(): void
     {
-        self::assertInstanceOf(Reference::class, Threads::client('access-token'));
+        self::assertInstanceOf(Client::class, Threads::client('access-token'));
     }
 
     public function testThreadsFactory(): void
     {
         self::assertInstanceOf(
-            Reference::class,
+            Client::class,
             Threads::factory()
             ->withHttpClient($this->createMock(ClientInterface::class))
             ->withConfig(new Config())
